+++ pkg/client/github/github.go
@@ -1,4 +1,4 @@
-package gitgrass
+package github
 
 import (
 	"fmt"
@@ -7,21 +7,17 @@ import (
 	"github.com/PuerkitoBio/goquery"
 )
 
-type HttpClient interface {
-	Get() (string, error)
+type GithubClient struct {
+	userName string
 }
 
-func NewClient(userName string) HttpClient {
-	return &GithubGrass{
+func NewClient(userName string) *GithubClient {
+	return &GithubClient{
 		userName: userName,
 	}
 }
 
-type GithubGrass struct {
-	userName string
-}
-
-func (g *GithubGrass) Get() (string, error) {
+func (g *GithubClient) Get() (string, error) {
 	res, err := http.Get(fmt.Sprintf("https://github.com/users/%s/contributions", g.userName))
 	if err != nil {
 		return "", err
